% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAaUF.r
\docType{data}
\name{db}
\alias{db}
\title{The 7 major currencies prices.}
\format{A matrix xts object with 2000 rows and 7 columns:
\describe{
  \item{price}{price, in US dollars}
  ...
}}
\source{
\url{www.activtrades.com}
}
\usage{
db
}
\description{
A dataset containing the 7 major currencies prices in xts format (2,000 rows).
}
\keyword{datasets}
