\name{n_trades_calculator}
\alias{n_trades_calculator}
\title{Trades number Calculator}
\usage{
n_trades_calculator (signals)
}
\description{
It calculate, according to a signal vector, the total trades occurred in it.
}
\arguments{
\item{signal}{an xts vector object containing Long/Short signals in -1/+1 format}

}

\examples{
}
