\name{pin_calc}
\alias{pin_calc}
\title{Shadow and Body extractor}
\usage{
pin_calc(ohlc)
}
\description{
This function, given an ohlc bar object, returns a list containing the up/low shadow and the body values.
}
\arguments{
\item{ohlc}{an open/high/low/close object}

}

\examples{

}
